+++ packages/commerce-sdk-react/CHANGELOG.md
@@ -1,3 +1,6 @@
+## v3.5.0-nightly-20250825080233 (Aug 25, 2025)
+## v3.5.0-dev (Aug 25, 2025)
+## v3.12.0-nightly-20250825080233 (Aug 25, 2025)
 ## v3.5.0-dev (Jul 22, 2025)
 - Add support for environment level base paths on /mobify routes [#2892](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2892)
 - Update USID expiry to match SLAS refresh token expiry[#2854](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2854)